% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoder_format.R
\name{add_split_comma}
\alias{add_split_comma}
\title{Pre-processes voter file by checking zipcode, and any special
characters or typos within the address.}
\usage{
add_split_comma(voter_file, address = "address", delimiter = "comma")
}
\arguments{
\item{voter_file}{A voter file containing the address of the voter.}

\item{address}{Either "single" or "split". single addresses are one
line address that include street, city, state, and zipcode on one line.}

\item{delimiter}{The type of delimiter (comma, hyphen, dash) that the
address parts are split into.}
}
\value{
The voter file with pre-processed format for each address
variable.
}
\description{
Pre-processes voter file by checking zipcode, and any special
characters or typos within the address.
}
