% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_utils.R
\name{betas_for_return}
\alias{betas_for_return}
\title{Manipulate precinct results to get betas as from ei_est_gen}
\usage{
betas_for_return(precinct_results, race_cand_pairs)
}
\description{
Manipulate precinct results to get betas as from ei_est_gen
}
