% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_utils.R
\name{check_args}
\alias{check_args}
\title{Check for missing essential arguments from an ei function}
\usage{
check_args(data, cand_cols, race_cols, totals_col, totals_null = TRUE)
}
\description{
Check for missing essential arguments from an ei function
}
