% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_preprocessing.R
\name{check_missing}
\alias{check_missing}
\title{Remove / Impute NAs in an EI dataset}
\usage{
check_missing(
  data,
  cand_cols,
  race_cols,
  totals_col,
  na_action = "drop",
  verbose = TRUE
)
}
\arguments{
\item{data}{A data.frame() object containing precinct-level turnout data by
race and candidate}

\item{cand_cols}{A character vector listing the column names for turnout for
each candidate}

\item{race_cols}{A character vector listing the column names for turnout by
race}

\item{totals_col}{The name of the column containing total votes cast in each
precinct}

\item{na_action}{A string indicating how to handle missing values in EI
columns. Possible values are "drop" and "mean". "drop" drops all observations
where variables are missing. "mean" imputes missing values as the mean of the
column}

\item{verbose}{A boolean indicating whether to return messages throughout the
function.}
}
\description{
Remove / Impute NAs in an EI dataset
}
