% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_preprocessing.R
\name{check_race_diffs}
\alias{check_race_diffs}
\title{Internal function that checks for adequate closeness between sums
of race columns and provided vote totals.}
\usage{
check_race_diffs(vote_sums, provided_totals, max_dev, avg_dev)
}
\arguments{
\item{vote_sums}{A numeric vector containing the sums of votes by each race/ethnicity}

\item{provided_totals}{A numeric vector containing the totals provided to the clean_race function}

\item{max_dev}{A flaoting point object passed from clean_race}

\item{avg_dev}{A floating point object passed from clean_race}
}
\value{
A numeric vector 0, 1, or 2.
}
\description{
This function checks conditions inputted to clean_race by the user
}
\details{
When this function returns 0, votes do not sum close enough 
to the provided vote totals. 

When this function returns 1, votes sum  marginally away 
from the provided totals. Restandardization is appropriate
 
When this function returns 2, votes sum exactly to vote
totals
}
