% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_preprocessing.R
\name{clean_race}
\alias{clean_race}
\title{Converts raw vote totals from different race/ethnicies across precincts 
into proportions, checking for problematic differences between known
vote totals and sums across race/ethnicities.}
\usage{
clean_race(
  data,
  cols,
  totals_col = NULL,
  max_dev = 0.1,
  avg_dev = 0.025,
  verbose = TRUE,
  diagnostic = FALSE
)
}
\arguments{
\item{data}{A dataframe of election results, where each row represents a precinct or geographic voting unit.}

\item{cols}{A character vector with the names of the columns indicating total votes cast by each race}

\item{totals_col}{A character string with the name of the total vote count column in the data. If null, total votes are computed within the function.}

\item{max_dev}{A numeric type object indicating the maximum allowable deviation of a precinct's vote sum from the totals in totals_col.}

\item{avg_dev}{A numeric type object indicating the maximum average deviation difference of all precints' vote sums from the totals in totals_col.}

\item{verbose}{A boolean indicating whether to print warnings and messages.}

\item{diagnostic}{A boolean. When true, an extra column of booleans is returned indicating whether each row had a deviation from totals_col}
}
\value{
A dataframe with proportions corresponding to the turnout of each race/ethnicity group.
}
\description{
If race turnout columns sum row-wise to equal vote_totals, they are 
returned as proportions. 
 
If race turnout columns sum row-wise to sufficiently close to 
vote_totals, they are returned as proportions of the sums.
 
If race turnout columns sum row-wise exceedingly far from 
vote_totals, the function stops and returns an error message.
}
