% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoder_format.R
\name{concat_final_address}
\alias{concat_final_address}
\title{This function concatenate the final address}
\usage{
concat_final_address(
  voter_file,
  street_address = "street_address",
  city = "city",
  state = "state",
  zipcode = "zipcode"
)
}
\arguments{
\item{voter_file}{A voter file containing the address of the voter.}

\item{street_address}{The street number and street name of the voters
address. Ex. 1442 Market Street}

\item{city}{The name of the city that the voter lives in.}

\item{state}{The state (based on the United States 50 states) that
the voter lives in.}

\item{zipcode}{The United States Postal Service (USPS) postal code.}
}
\value{
The voter file with pre-processed format for each address variable.
}
\description{
This function concatenate the final address
}
