% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoder_format.R
\name{concat_streetname}
\alias{concat_streetname}
\title{Pre-processes voter file by checking zipcode, and any special characters
or typos within the address.}
\usage{
concat_streetname(
  voter_file,
  street_number = "street_number",
  street_name = "street_name",
  street_suffix = "street_suffix"
)
}
\arguments{
\item{voter_file}{A voter file containing the address of the voter.}

\item{street_number}{The number attached to the street name. Ex. 1442}

\item{street_name}{The name of the place in which a voter lives.
Ex. Market Street}

\item{street_suffix}{A directional abbreviation such as NE for northeast or
SW for southwest.}
}
\value{
The voter file with pre-processed format for each address variable.
}
\description{
Pre-processes voter file by checking zipcode, and any special characters
or typos within the address.
}
