% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiCompare-package.R
\docType{data}
\name{cor_06}
\alias{cor_06}
\title{Corona 2006}
\format{
A data frame with 47 observations on the following 8 variables.
\describe{ \item{list("precinct")}{a numeric vector}
\item{list("totvote")}{a numeric vector} \item{list("pct_latino")}{a numeric
vector} \item{list("pct_other")}{a numeric vector}
\item{list("pct_breitenbucher")}{a numeric vector}
\item{list("pct_montanez")}{a numeric vector} \item{list("pct_spiegel")}{a
numeric vector} \item{list("pct_skipworth")}{a numeric vector} }
}
\description{
Precinct vote data from a Corona, CA 2006 election
}
\examples{

data(cor_06)
# Look at data
head(cor_06)
str(cor_06)

}
\references{
Riverside County, CA board of elections
}
\keyword{datasets}
