% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiCompare-package.R
\docType{data}
\name{corona}
\alias{corona}
\title{Corona Precinct Racial Bloc Voting Data}
\format{
A data frame with 46 observations on the following 12 variables.
\describe{ \item{list("precinct")}{a numeric vector}
\item{list("totvote")}{a numeric vector} \item{list("pct_husted")}{a numeric
vector} \item{list("pct_spiegel")}{a numeric vector}
\item{list("pct_ruth")}{a numeric vector} \item{list("pct_button")}{a
numeric vector} \item{list("pct_montanez")}{a numeric vector}
\item{list("pct_fox")}{a numeric vector} \item{list("pct_hisp")}{a numeric
vector} \item{list("pct_asian")}{a numeric vector}
\item{list("pct_white")}{a numeric vector} \item{list("pct_non_lat")}{a
numeric vector} }
}
\description{
Data taken from a 2014 California election, with precinct results and racial
demographics for Corona, CA precincts
}
\examples{

data(corona)
head(corona)
str(corona)

}
\references{
Riverside County, CA board of elections
}
\keyword{datasets}
