% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voter_file_utils.R
\name{dedupe_voter_file}
\alias{dedupe_voter_file}
\title{De-duplicates a voter file.}
\usage{
dedupe_voter_file(voter_file, voter_id = "voter_id")
}
\arguments{
\item{voter_file}{The voter file, as a data frame or tibble.}

\item{voter_id}{The column denoting the voter ID.}
}
\value{
The voter file with duplicates removed.
}
\description{
Currently, this function removes all but the latest entries in a voter file
according to voter ID. This assumes the voter file is sorted by voter ID
chronologically.
}
\details{
This function can be updated with more functionality to handle edge cases.
}
