% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_sim.R
\name{.createR}
\alias{.createR}
\title{This function is pulled directly from the ei package to enable a modified
version of ei.sim()}
\usage{
.createR(sub, Rfun, bb, bw, sb, sw, rho, x, numb, numw)
}
\arguments{
\item{sub}{passed from ei_sim()}

\item{Rfun}{passed from ei_sim()}

\item{bb}{passed from ei_sim()}

\item{bw}{passed from ei_sim()}

\item{sb}{passed from ei_sim()}

\item{sw}{passed from ei_sim()}

\item{rho}{passed from ei_sim()}

\item{x}{passed from ei_sim()}

\item{numb}{passed from ei_sim()}

\item{numw}{passed from ei_sim()}
}
\description{
This function is pulled directly from the ei package to enable a modified
version of ei.sim()
}
