% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_est_gen.R
\name{ei_est_gen}
\alias{ei_est_gen}
\title{Iterative EI Estimation}
\usage{
ei_est_gen(
  data,
  cand_cols,
  race_cols,
  totals_col,
  rho = 10,
  table_names,
  sample = 1000,
  tomog = FALSE,
  density_plot = FALSE,
  beta_yes = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{A data.frame object containing turnout information by 
candidate and race}

\item{cand_cols}{A character vector of the names of columns containing
data on turnout for each candidate}

\item{race_cols}{A character vector of the names of columns containing
data on turnout by race group.}

\item{totals_col}{The name of the column containing vote totals for each
row}

\item{rho}{The rho parameter for ei() estimate, defaults to 10, numeric}

\item{table_names}{A character vector of table names with same length as
race_cols. Used for formatting output. If only one racial group, must
provide "Pct. Other" as second element of vector}

\item{sample}{The number of samples used for EI calculation, default = 1000}

\item{tomog}{A logical to display tomography plot. If true will will save pdf
plot to working directory. Default is FALSE}

\item{density_plot}{A logical to display density plot of betab and betaw. If
true will save pdf plot to working directory. Default is FALSE}

\item{beta_yes}{A logical to export betas (b, w) in list object in addition to
table of results. Default is FALSE}

\item{verbose}{A logical to print messages, ei summaries as iterations progress.}

\item{\dots}{Arguments passed onto ei() function}
}
\value{
Data frame/table object containing EI individually estimated
results. If beta_yes=TRUE, two list items, first the data frame table of
results, second dataframe of betas themselves.
}
\description{
Iteratively fits EI models for candidates and racial/ethnic groups
}
\note{
If this results in an error, "Error in .subset2(x, i, exact = exact) :
invalid subscript type 'list'", just rerun the algorithm again.
}
\examples{

# TOY DATA EXAMPLE
canda <- c(.1, .09, .85, .9, .92)
candb <- 1 - canda
white <- c(.8, .9, .10, .08, .11)
black <- 1 - white
total <- c(30, 80, 70, 20, 29)
toy <- data.frame(canda, candb, white, black, total)

# CREATE VECTORS
cand_cols <- c("canda")
race_cols <- c("black") # only use one group for example
table_names <- c("EI: PCT Black", "EI: PCT White")

# RUN ei_est_gen()
# KEEP DATA TO JUST ONE ROW FOR EXAMPLE (time) ONLY!
ei_est_gen(
  data = toy[c(1, 3, 5), ],
  cand_cols = cands, 
  race_cols = race_group, 
  totals_col = "total",
  table_names = table_names,
  sample = 100
)
\donttest{
# WARNING -- May take a little while to execute
# Load Package Data
data(corona)
# Create Character Vectors
cands <- c("pct_husted", "pct_spiegel", "pct_ruth", "pct_button", "pct_montanez", "pct_fox")
race_group3 <- c("pct_hisp", "pct_asian", "pct_white")
table_names <- c("EI: Pct Hisp", "EI: Pct Asian", "EI: Pct White")

# Run ei_est_gen function
results <- ei_est_gen(
  data = corona, 
  cand_cols = cands, 
  race_group = race_group3,
  total = "totvote", 
  table_names = table_names
)

results
# Run ei_est_gen function; Exporting betas into data frame
results_w_betas <- ei_est_gen(
  data = corona, 
  cand_cols = cands, 
  race_group = race_group3,
  totals_col = "totvote", 
  table_names = table_names, 
  beta_yes = TRUE
)

res1 <- results_w_betas[[1]] # table of mean estimates
res1
res2 <- results_w_betas[[2]] # betas of estimates for each precinct
}

}
\references{
eiPack. Gary King (1997). A Solution to the Ecological Inference
Problem. Princeton: Princeton University Press.
}
\author{
Loren Collingwood <loren.collingwood@ucr.edu>
}
