% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_iter.R
\name{ei_iter}
\alias{ei_iter}
\title{Iterative EI Estimation}
\usage{
ei_iter(
  data,
  cand_cols,
  race_cols,
  totals_col,
  erho = 10,
  sample = 1000,
  plots = FALSE,
  betas = FALSE,
  ...
)
}
\arguments{
\item{data}{A data.frame() object containing precinct-level turnout data by race and candidate}

\item{cand_cols}{A character vector listing the column names for turnout for each candidate}

\item{race_cols}{A character vector listing the column names for turnout by race}

\item{totals_col}{The name of the column containing total votes cast in each precinct}

\item{sample}{The number of samples used in ei estimation. Defaulted to 1000}

\item{plots}{A boolean indicating whether or not to include density and tomography plots}

\item{betas}{A boolean to return precinct-level betas for each 2x2 ei}

\item{...}{Additional arguments passed directly to ei::ei()}

\item{rho}{A number passed directly to ei::ei(). Defaulted to 10}
}
\value{
dataframe of results from iterative ei
}
\description{
Iterative EI Estimation
}
