% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_rc_congruence.R
\name{ei_rc_congruence}
\alias{ei_rc_congruence}
\title{Congruence for 2x2}
\usage{
ei_rc_congruence(ei_rc_table, cand_race, group_race)
}
\arguments{
\item{ei_rc_table}{Object produced from ei_rc_good_table(), where
include_good=F, of class ei_compare}

\item{cand_race}{Numeric vector indicating race of the candidates in order
they show up in table rownames, where 1=Latino; 2=Black; 3=Asian;
4=White/Non}

\item{group_race}{Numeric vector, taking similar values as cand_race where
1=Latino; 2=Black; 3=Asian; 4=White/Non}
}
\value{
Table of congruence scores
}
\description{
Calculates congruence scores between EI and RxC for the 2x2 Scenario
}
\author{
Loren Collingwood <loren.collingwood@ucr.edu>, Matt Barreto
<barretom@ucla.edu>
}
