% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiCompare-package.R
\docType{package}
\name{eiCompare-package}
\alias{eiCompare-package}
\alias{eiCompare}
\title{Compares EI, Goodman, RxC Estimates}
\description{
Compares estimates from three ecological inferences routines, based on King
et. al.'s approach.
}
\details{
See demo(demo, "eiCompare") for examples on how to use code
}
\references{
Gary King (1997). A Solution to the Ecological Inference
Problem. Princeton: Princeton University Press. Lau, Olivia, Ryan Moore, and
Michael Kellerman. eiPack: Ecological Inference and Higher-Dimension Data
Management
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RPVote/eiCompare}
}

}
\author{
Loren Collingwood

Maintainer: Loren Collingwood <lcollingwood@unm.edu>
}
\keyword{package}
