% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{empty_ei_df}
\alias{empty_ei_df}
\title{Create a dataframe with NA values for racial and candidate counts.}
\usage{
empty_ei_df(ncand = 2, nrace = 2, nrow = 2)
}
\arguments{
\item{ncand}{The number of candidates to include}

\item{nrace}{The number of race/ethnicitiesto include}

\item{nrow}{The number of rows for the dataframe}
}
\value{
A dataframe with columns for each candidate and race, all with NAs
}
\description{
Create a dataframe with NA values for racial and candidate counts.
}
