% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ersd_maps}
\alias{ersd_maps}
\title{East Ramapo School District Proposed Maps}
\format{
A data frame with 46 observations on the following 12 variables:
\describe{
 \item{WARD}{Precinct ID number.}
 \item{TOT_CVAP}{The total vote, per precinct.}
 \item{WHI_CVAP}{Percent of vote for Husted.}
 \item{BLA_CVAP}{Percent of vote for Spiegel.}
 \item{HIS_CVAP}{Percent of vote for Ruth.}
 \item{ASI_CVAP}{Percent of vote for Button.}
 \item{MIN_AGG_FRAC}{Percent of vote for Montanez.}
 \item{geometry}{Percent of vote for Fox.}
}
}
\source{
East Ramapo School District
}
\usage{
data(ersd_maps)
}
\description{
This dataset contains proposed maps and CVAP totals for East Ramapo School
District.
}
\keyword{datasets}
