% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_rpvDF}
\alias{example_rpvDF}
\title{Example RPV analysis results in Washington State}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 72 rows and 13 columns.
}
\usage{
example_rpvDF
}
\description{
Example RPV analysis results in Washington State
}
\keyword{datasets}
