% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_utils.R
\name{get_ei_est_gen_results}
\alias{get_ei_est_gen_results}
\title{Convert ei objects into ei_est_gen-style results tables}
\usage{
get_ei_est_gen_results(ei_objects, race_cand_pairs, betas)
}
\arguments{
\item{ei_objects}{A list of ei objects}

\item{race_cand_pairs}{Passed from ei_iter}

\item{betas}{Passed from ei_iter}
}
\description{
Convert ei objects into ei_est_gen-style results tables
}
\author{
Ari Decter-Frain <agd75@cornell.edu>
}
