% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_utils.R
\name{get_ei_iter_se}
\alias{get_ei_iter_se}
\title{Get 2x2 ei standard errors from ei object
Works according to the aggregate formula in King, 1997, section 8.3}
\usage{
get_ei_iter_se(aggs)
}
\arguments{
\item{aggs}{A dataframe of aggregate value draws, taken from eiread()}
}
\description{
Get 2x2 ei standard errors from ei object
Works according to the aggregate formula in King, 1997, section 8.3
}
\author{
Ari Decter-Frain <agd75@cornell.edu>
}
