% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_utils.R
\name{get_md_bayes_gen_output}
\alias{get_md_bayes_gen_output}
\title{Get md_bayes_gen() output from ei_rxc() output}
\usage{
get_md_bayes_gen_output(results_table, race_cols)
}
\arguments{
\item{results_table}{A results table from}

\item{race_cols}{Character vector of candidate race names, passed from
ei_rxc}
}
\description{
Get md_bayes_gen() output from ei_rxc() output
}
\author{
Ari Decter-Frain <agd75@cornell.edu>
}
