% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surname_utils.R
\name{get_multi_barreled_surnames}
\alias{get_multi_barreled_surnames}
\title{Gets multi-barreled surnames from a voter file.}
\usage{
get_multi_barreled_surnames(
  voter_file,
  surname_col = "last_name",
  regex = "[ -]+"
)
}
\arguments{
\item{voter_file}{The voter file, with each row consisting of a voter.}

\item{surname_col}{A string denoting the surname column.}

\item{regex}{A string denoting the regular expression to use for denoting the
the special characters.}
}
\value{
A dataframe of voters whose surnames are multi-barreled.
}
\description{
A multi-barreled surname is one containing a dash or a space. This function
finds all multi-barreled surnames in a voter file.
}
