% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EI_utils.R
\name{get_results_table}
\alias{get_results_table}
\title{Get results dataframe from a list of results}
\usage{
get_results_table(
  district_results,
  cand_col = cand_cols,
  race_col = race_cols,
  n_cand = n_cands,
  n_race = n_races,
  n_iter = n_iters
)
}
\description{
Get results dataframe from a list of results
}
