% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surname_utils.R
\name{get_special_character_surnames}
\alias{get_special_character_surnames}
\title{Gets surnames containing special characters.}
\usage{
get_special_character_surnames(
  voter_file,
  surname_col = "last_name",
  regex = "[^A-Za-z]"
)
}
\arguments{
\item{voter_file}{The voter file, with each row consisting of a voter.}

\item{surname_col}{A string denoting the surname column.}

\item{regex}{A string denoting the regular expression to use for querying the
the special characters.}
}
\value{
A dataframe of voters whose surname has special characters.
}
\description{
Returns a subsetted voter file whose rows consist of voters that have
special characters in their last name.
}
