% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surname_utils.R
\name{get_word_count}
\alias{get_word_count}
\title{Counts the number of words per row in the column of a dataframe.}
\usage{
get_word_count(voter_file, surname_col = "last_name", regex = "[ -]+")
}
\arguments{
\item{voter_file}{The voter file, with each row consisting of a voter.}

\item{surname_col}{A string denoting the surname column.}

\item{regex}{A string denoting the regular expression to use for querying the
the word count.}
}
\value{
A vector of word counts.
}
\description{
A "word" is defined as a string of alphabetical characters separated by
either spaces or dashes (but not other special characters).
}
