% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwin_fulton_shape}
\alias{gwin_fulton_shape}
\title{Shape file information for Gwinnett and Fulton counties in Georgia}
\format{
A data frame with 68 rows and 17 columns
\describe{
\item{STATEFP10}{the 2010 FIPS code for the state geograhic level}
\item{COUNTYFP10}{the 2010 FIPS code for the county geographic level}
\item{TRACTCE10}{the 2010 FIPS code for the tract geographic level}
\item{BLOCKCE10}{the 2010 FIPS code for the block geographic level}
\item{GEOID10}{the 2010 FIPS code for Census block identifier.
a concatenation of 2010 Census state FIPS code, 2010 Census
county FIPS code, 2010 Census tract code, and 2010 Census block number}
\item{NAME10}{Census block identifier; a concatenation of 2010
Census state FIPS code, 2010 Census county FIPS
code, 2010 Census tract code, and 2010 Census
block number}
\item{MTFCC10}{MAF/TIGER feature class code (G5040)}
\item{UR10}{2010 Census urban/rural indicator}
\item{UACE10}{2010 Census urban area code}
\item{UATYPE}{2010 Census urban area type}
\item{FUNCSTAT10}{2010 Census functional status}
\item{ALAND10}{2010 Census land area}
\item{AWATER10}{2010 Census water area}
\item{INTPTLAT10}{2010 Census latitude of the internal point}
\item{INTPTLON10}{2010 Census longitude of the internal point}
\item{geometry}{latitude and longitude coordinates}
\item{COUNTYFP}{the FIPS code for county}
\item{STATEFP}{the FIPS code for state}
}
}
\usage{
data(gwin_fulton_shape)
}
\description{
This dataset contains results for shape file FIPS codes and geometrie
using the tigris package from the US Census Bureau. The values correspond
to information about the multipolygon geometry and fips code values for
state, county, tract, and block geographies.
}
\keyword{datasets}
