% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwinnett_ei}
\alias{gwinnett_ei}
\title{Stylized dataset of election results and turnout by race in Gwinnett county,
2018 Georgia gubernatorial election.}
\format{
A data frame with 157 rows and 5 columns
\describe{
\item{kemp}{Proportion of votes cast for Republican candidate Brian Kemp}
\item{abrams}{Proportion of votes cast for Democratic candidate Stacey Abrams}
\item{white}{Proportion of voters self-reporting as white}
\item{nonwhite}{Proportion of voters self-reporting as non-white}
\item{turnout}{Count of voter turnout}
}
}
\usage{
data(gwinnett_ei)
}
\description{
Stylized dataset of election results and turnout by race in Gwinnett county,
2018 Georgia gubernatorial election.
}
\keyword{datasets}
