% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiCompare-package.R
\docType{data}
\name{lac_10}
\alias{lac_10}
\title{Los Angeles County Data}
\format{
A data frame with 4980 observations on the following 10 variables.
\describe{ \item{list("precinct")}{a character vector}
\item{list("tot_reg")}{a numeric vector} \item{list("i_jones")}{a numeric
vector} \item{list("i_delatore")}{a numeric vector}
\item{list("votescast")}{a numeric vector} \item{list("lat_voters")}{a
numeric vector} \item{list("pct_latino")}{a numeric vector}
\item{list("pct_delatorre")}{a numeric vector} \item{list("pct_jones")}{a
numeric vector} \item{list("pct_other")}{a numeric vector} }
}
\source{
Los Angeles county
}
\description{
Los Angeles County precinct dataset from 2010 election
}
\examples{

data(lac_10)
head(lac_10)
str(lac_10)
}
\keyword{datasets}
