% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_sim.R
\name{like}
\alias{like}
\title{Likelihood function for ecological inference.}
\usage{
like(param, y, x, n, Zb, Zw, numb, erho, esigma, ebeta, ealphab, ealphaw, Rfun)
}
\description{
This function is pulled directly from the ei package to enable a modified
version of ei.sim()
}
