% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_shapefile.R
\name{map_shape_file}
\alias{map_shape_file}
\title{Function for making basic choropleth maps of shape_file
using the tigris package and ggplot}
\usage{
map_shape_file(
  shape_file,
  crs = "+proj=latlong +ellps=GRS80 +no_defs",
  title = "Title of the Shapefile"
)
}
\arguments{
\item{shape_file}{a shape file based on desired ecological
unit (i.e. state, county, block, tract)}

\item{crs}{the Coordinate reference system, default is
crs="+proj=latlong +ellps=GRS80 +no_defs"}

\item{title}{the tile of the map}
}
\value{
Plots of mapped ecological units desired shape
}
\description{
Function for making basic choropleth maps of shape_file
using the tigris package and ggplot
}
\author{
Loren Collingwood <loren.collingwood@ucr.edu>

Juandalyn Burke <jcburke@uw.edu>
}
