% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_shape_points.R
\name{map_shape_points}
\alias{map_shape_points}
\title{Function for making basic choropleth maps of longitude and latitude
points using the tigris package and ggplot}
\usage{
map_shape_points(
  voter_file,
  shape_file,
  crs = "+proj=longlat +ellps=GRS80",
  title = "title"
)
}
\arguments{
\item{voter_file}{a voter file with latitude, longitude, or
latitude/longitude and other geographic data.}

\item{shape_file}{a shape file based on desired ecological
unit (i.e. state, county, block, tract)}

\item{crs}{the Coordinate reference system, default is
crs="+proj=latlong +ellps=GRS80 +no_defs"}

\item{title}{the tile of the map}
}
\value{
Plots of mapped ecological units desired and voter 
latitude and longitudes
}
\description{
Function for making basic choropleth maps of longitude and latitude
points using the tigris package and ggplot
}
\author{
Loren Collingwood <loren.collingwood@ucr.edu>

Juandalyn Burke <jcburke@uw.edu>
}
