% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_bayes_gen.R
\name{md_bayes_gen}
\alias{md_bayes_gen}
\title{MD Bayes Generalize}
\usage{
md_bayes_gen(
  data,
  cand_cols,
  race_cols,
  totals_col = NULL,
  ntunes = 10,
  totaldraws = 10000,
  seed = 12345,
  sample = 1000,
  thin = 100,
  burnin = 10000,
  ret.mcmc = TRUE,
  ci = c(0.025, 0.975),
  ci_true = TRUE,
  produce_draws = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{A data.frame object containing turnout information by 
candidate and race}

\item{cand_cols}{A character vector of the names of columns containing
data on turnout for each candidate}

\item{race_cols}{A character vector of the names of columns containing
data on turnout by race group.}

\item{totals_col}{The name of the column containing vote totals for each
row. If NULL, assumes columns are totals, not pct.}

\item{ntunes}{Numeric. How much to tune tuneMD. Default = 10}

\item{totaldraws}{Numeric. Number of total draws from MD. Default = 10000}

\item{seed}{Numeric. Default = 12345}

\item{sample}{Numeric. Default = 10000}

\item{thin}{Numeric. Default = 10}

\item{burnin}{Numeric. Default = 10000}

\item{ci}{numeric vector of credible interval (low/high), default is 95
percent= c(0.025, 0.975)}

\item{ci_true}{Logical, default = TRUE. Include credible intervals in
reported results.}

\item{produce_draws}{Logical, default is FALSE. Produces two-item list of
table and md.bayes() mcmc draws (for additional testing and analysis)}

\item{...}{Additional arguments passed to tuneMD() and ei.MD.bayes()}

\item{ret_mcmc}{Logical. Default = TRUE}
}
\value{
List object of length 1 (when produce_draws=FALSE). List object of
length 2 (when produce_draws=TRUE). First item is list of race x candidate
tabular results, with mean, SE, and credible intervals. Second item is mcmc
draws.
}
\description{
Tunes and estimates MD Bayes algorithm (ei.MD.bayes). This, combined with
md_bayes_table() produces tables of results compatible with EI table of
results.
}
\examples{


# TOY DATA EXAMPLE
canda <- c(.1, .09, .85, .9, .92)
candb <- 1 - canda
white <- c(.8, .9, .10, .08, .11)
black <- 1 - white
total <- c(30, 80, 70, 20, 29)
toy <- data.frame(canda, candb, white, black, total)

# only use one group for example
cand_cols <- c("canda", "candb)
race_cols <- c("black", "white)

# Then execute md_bayes_gen(); not run here due to time
md_bayes_gen(
  data = toy,
  cand_cols = cand_cols,
  race_cols = race_cols,
  totals_col = "total"
  ntunes = 1,
  thin = 1,
  totaldraws = 100,
  sample = 10,
  burnin = 1
)

# Add in mcmc drawings
drawings <- md_bayes_gen(
  dat = toy,
  form = form,
  total_yes = FALSE,
  ntunes = 1,
  thin = 1,
  totaldraws = 100,
  sample = 10,
  burnin = 1,
  produce_draws = TRUE
)
head(drawings$draws)
}
\references{
eiPack, King et. al. (http://gking.harvard.edu/eiR)
}
\author{
Loren Collingwood <loren.collingwood@ucr.edu>
}
