\name{mean_and_ci}
\alias{mean_and_ci}

\title{mean_and_ci
  
}
\description{Internal
  
}
\usage{mean_and_ci(cbind_dat, ci=c(.025, .975))
  
}

\arguments{
  \item{cbind_dat}{cbind object}
  \item{ci}{Credible intervals. Default: c(.025, .975)}
}

\value{
Mean and credible interval  
}
\author{
  Loren Collingwood <loren.collingwood@ucr.edu>
}

\examples{
  
  # EXAMPLE: NOT RUN #
  
}
