% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_plot_create.R
\name{od_plot_create}
\alias{od_plot_create}
\title{od_plot_create}
\usage{
od_plot_create(race, cand_pair, dens_data, out, plot_path = "", cand_colors)
}
\arguments{
\item{race}{Racial demographic of interest}

\item{cand_pair}{All possible candidate pairing combinations}

\item{dens_data}{Beta values long for each race and candidate pair}

\item{out}{Summary table from overlay_density_plot for every race candidate
pair}

\item{plot_path}{Path to save plots}

\item{cand_colors}{Colors for every candidate}
}
\value{
Comparison density plots

overlay density plot comparing candidates for votes by race
}
\description{
Internal
}
\author{
Loren Collingwood <loren.collingwood@ucr.edu>

Hikari Murayama
}
