% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_density_plot.R
\name{overlay_density_plot}
\alias{overlay_density_plot}
\title{overlay_density_plot}
\usage{
overlay_density_plot(
  agg_betas,
  results_table,
  race_cols,
  cand_cols,
  plot_path,
  ei_type
)
}
\arguments{
\item{agg_betas}{Output for RxC and iterative ei}

\item{results_table}{Summary table for candidate race pair means and se's}

\item{race_cols}{A character vector listing the column names for turnout by race}

\item{cand_cols}{A character vector listing the column names for turnout for each candidate}

\item{plot_path}{Path to save}

\item{ei_type}{Specify whether the data comes from iterative ei ("ei") or rxc ("rxc")}
}
\value{
Prep and run density plot creation iteratively
}
\description{
Internal
}
\author{
Loren Collingwood <loren.collingwood@ucr.edu>

Hikari Murayama
}
