% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.eiCompare.R
\name{plot.eiCompare}
\alias{plot.eiCompare}
\title{Print a plot comparing the predictions of EI outputs.}
\usage{
\method{plot}{eiCompare}(x, ...)
}
\arguments{
\item{x}{An eiCompare object, outputted from ei_iter() or ei_rxc().}

\item{...}{Additional eiCompare objects to summarize.}
}
\value{
A ggplot comparing eiCompare objects.
}
\description{
The output of this plot consists of error bars containing the mean for each
candidate, racial group, and eiCompare object. Error bars represent one
standard deviation from the mean of the posterior sampling distribution.
}
