% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bivariate.R
\name{plot_bivariate}
\alias{plot_bivariate}
\title{Plot bivariate relationships}
\usage{
plot_bivariate(data, cand_cols, race_cols)
}
\arguments{
\item{data}{A data.frame() object containing precinct-level turnout data by
race and candidate}

\item{cand_cols}{A character vector listing the column names for turnout for
each candidate}

\item{race_cols}{A character vector listing the column names for turnout by
race}
}
\value{
ggplot object with bivariate plots faceted by
}
\description{
Plot bivariate relationships
}
