% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eic.R
\name{plot_eic}
\alias{plot_eic}
\title{Plot eiCompare objects for comparison}
\usage{
plot_eic(eic_objects)
}
\arguments{
\item{eic_objects}{A named list of eiCompare objects to be compared in plot.
Names will be used to label each object in the plot legend.}
}
\value{
A ggplot2 object
}
\description{
Plot eiCompare objects for comparison
}
