% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surname_utils.R
\name{predict_race_double_barreled}
\alias{predict_race_double_barreled}
\title{Predicts, for one row in a voter file, the probability of a voter having a
certain race by averaging over each "barrel" of the surname.}
\usage{
predict_race_double_barreled(
  df,
  surname_col = "last_name",
  surname_only = TRUE,
  pattern = "[ -]+"
)
}
\arguments{
\item{surname_col}{A string denoting the surname column.}

\item{surname_only}{Whether to obtain probabilities for surnames only.}

\item{pattern}{What pattern to split surnames on. By default, surnames are
split on a space(s), which assumes hyphens have already been removed.}

\item{voter_file}{The voter file, with each row consisting of a voter.}
}
\value{
A vector of probabilities for each surname.
}
\description{
Predicts, for one row in a voter file, the probability of a voter having a
certain race by averaging over each "barrel" of the surname.
}
