% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_descriptives.R
\name{race_cand_cors}
\alias{race_cand_cors}
\title{Table of bivariate correlations}
\usage{
race_cand_cors(data, cand_cols, race_cols)
}
\arguments{
\item{data}{A data.frame() object containing precinct-level turnout data by
race and candidate}

\item{cand_cols}{A character vector listing the column names for turnout for
each candidate}

\item{race_cols}{A character vector listing the column names for turnout by
race}
}
\description{
Table of bivariate correlations
}
