% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ramapo2018}
\alias{ramapo2018}
\title{East Ramapo School District 2018 Voter File}
\format{
A data frame with 9401 observations on the following 7 variables:
\describe{
 \item{voter_id}{The voter ID, recoded to protect voter privacy.}
 \item{last_name}{The surname of the voter.}
 \item{ward}{The ward, as a character, that the voter is located in.}
 \item{state}{FIPS code of state for voter.}
 \item{county}{FIPS code of Census county for voter.}
 \item{tract}{FIPS code of Census tract for voter.}
 \item{block}{FIPS code of Census block for voter.}
}
}
\source{
East Ramapo School District
}
\usage{
data(ramapo2018)
}
\description{
This dataset contains a subset of the voter file for voters in East Ramapo
School District, in 2018. This file has been modified to protect the privacy
of the voters. The voter IDs have been replaced, surnames have been replaced
with "similar" surnames, and the file comes already geocoded into Census
block and ward.
}
\keyword{datasets}
