% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EI_utils.R
\name{remove_nas}
\alias{remove_nas}
\title{Remove missing values from dataset and return warning if any removed}
\usage{
remove_nas(data)
}
\description{
Remove missing values from dataset and return warning if any removed
}
