% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpv_density.R
\name{rpv_density}
\alias{rpv_density}
\title{rpv_density}
\usage{
rpv_density(agg_betas, plot_path = NULL)
}
\arguments{
\item{agg_betas}{Aggregated beta values}

\item{plot_path}{Path to save. If NULL, plot is not saved.}
}
\value{
Return density for every race/candidate pair for Bb-Bw
}
\description{
rpv_density
}
\examples{


# EXAMPLE: NOT RUN #
}
\author{
Loren Collingwood <loren.collingwood@ucr.edu>

Hikari Murayama
}
