% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_utils.R
\name{rxc_formula}
\alias{rxc_formula}
\title{Make rxc formula}
\usage{
rxc_formula(cand_cols, race_cols)
}
\arguments{
\item{cand_cols}{Character vector of candidate column names, passed from
ei_rxc}

\item{race_cols}{Character vector of candidate race names, passed from
ei_rxc}
}
\description{
Make rxc formula
}
\author{
Ari Decter-Frain <agd75@cornell.edu>
}
