% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{south_carolina}
\alias{south_carolina}
\title{Example election and demographic data from South Carolina 2020 General Elections}
\format{
An object of class \code{data.frame} with 750 rows and 42 columns.
}
\usage{
south_carolina
}
\description{
Example election and demographic data from South Carolina 2020 General Elections
}
\keyword{datasets}
