% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoder_format.R
\name{split_add_nocommas}
\alias{split_add_nocommas}
\title{Pre-processes voter file by checking zipcode, and any special
characters or typos within the address.}
\usage{
split_add_nocommas(voter_file, address = "address", delimiter = "space")
}
\arguments{
\item{voter_file}{A voter file containing the address of the voter.}

\item{address}{The voter address in the structure of street
number-street name-city-state-zipcode}

\item{delimiter}{The symbol/character used to parse address."space"
is used to indicate that the address is parsed by spaces.}
}
\value{
The voter file with pre-processed format for each address
variable.
}
\description{
Pre-processes voter file by checking zipcode, and any special
characters or typos within the address.
}
\author{
Loren Collingwood <loren.collingwood@ucr.edu>

Juandalyn Burke <jcburke@uw.edu>
}
