% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_preprocessing.R
\name{sum_over_cols}
\alias{sum_over_cols}
\title{Sum row-wise over columns in a dataframe}
\usage{
sum_over_cols(data, cols)
}
\arguments{
\item{data}{A data.frame() object containing precinct-level turnout data by
race and candidate}

\item{cols}{A set of columns to sum over. Typically, enter cand_cols or
race_cols here.}
}
\value{
A vector of row-wise sums across the column vector entered as
argument.
}
\description{
Simple wrapper of rowSums for checking row sums of race, candidate columns
}
