% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.eiCompare.R
\name{summary.eiCompare}
\alias{summary.eiCompare}
\title{Print a summary of an eiCompare object}
\usage{
\method{summary}{eiCompare}(object, ...)
}
\arguments{
\item{object}{An eiCompare object, outputted from ei_iter() or ei_rxc()}

\item{...}{Additional eiCompare objects to summarize}
}
\value{
A nicely formatted dataframe for printing results
}
\description{
Print a summary of an eiCompare object
}
