% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.eic.R
\name{summary_eic}
\alias{summary_eic}
\title{Print a summary of an eiCompare object}
\usage{
summary_eic(eic_object)
}
\arguments{
\item{eic_object}{An eiCompare object, outputted from ei_iter() or ei_rxc()}
}
\value{
A nicely formatted dataframe for printing results
}
\description{
Print a summary of an eiCompare object
}
