% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surname_utils.R
\name{surname_match}
\alias{surname_match}
\title{Determines which surnames match to the Census list.}
\usage{
surname_match(voter_file, surname_col = "last_name", strip_special = FALSE)
}
\arguments{
\item{voter_file}{The voter file, with each row consisting of a voter.}

\item{surname_col}{A string denoting the surname column.}

\item{strip_special}{Whether to strip special characters before matching in
the surname database.}
}
\value{
A vector of logicals denoting a match or not.
}
\description{
Determines which surnames match to the Census list.
}
