% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surname_utils.R
\name{surname_summary}
\alias{surname_summary}
\title{Briefly summarizes the surnames in a voter file.}
\usage{
surname_summary(voter_file, surname_col)
}
\arguments{
\item{voter_file}{The voter file, with each row consisting of a voter.}

\item{surname_col}{A string denoting the surname column.}
}
\description{
Briefly summarizes the surnames in a voter file.
}
