\name{unit_comb_extract}
\alias{unit_comb_extract}

\title{Extract relevant geographic units from 15-digit FIPS code
  
}
\description{Extract relevant geographic units from 15-digit FIPS code
  
}
\usage{
  unit_comb_extract(fips)
}

\arguments{
  \item{fips}{Character vector. 15-digit FIPS code}
}

\value{List with separate datasets of length fips, including values for fips, state, county, tract, block, state-county, and state-county-tract.
}

\author{
  Loren Collingwood <loren.collingwood@ucr.edu>
}

\examples{
  

}
