% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoder_format.R
\name{zip_hyphen}
\alias{zip_hyphen}
\title{Pre-processes voter file by checking zipcode, and any special
characters or typos within the address.}
\usage{
zip_hyphen(voter_file, voter_id = "registration_number", zipcode = "zipcode")
}
\arguments{
\item{voter_file}{A voter file containing the address of the voter.}

\item{voter_id}{The unique identifier linked to the voter.}

\item{zipcode}{The United States Postal Service (USPS) postal code.}
}
\value{
The voter file with pre-processed format for each address
variable.
}
\description{
Pre-processes voter file by checking zipcode, and any special
characters or typos within the address.
}
