% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_utils.R
\name{betas_for_return}
\alias{betas_for_return}
\title{Manipulate precinct results to get betas as from ei_est_gen}
\usage{
betas_for_return(precinct_results, race_cand_pairs)
}
\arguments{
\item{precinct_results}{A list of betas from ei_iter()}

\item{race_cand_pairs}{The set of race/candidate pairs tested in ei_iter}
}
\description{
Manipulate precinct results to get betas as from ei_est_gen
}
\author{
Ari Decter-Frain <agd75@cornell.edu>
}
