% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_utils.R
\name{check_args}
\alias{check_args}
\title{Check for missing essential arguments from an ei function}
\usage{
check_args(data, cand_cols, race_cols, totals_col, totals_null = FALSE)
}
\arguments{
\item{data}{A dataframe upon which EI is to be performed}

\item{cand_cols}{A column of candidate names passed from ei functions}

\item{race_cols}{A column of race names passed from ei functions}

\item{totals_col}{The name of a column passed from ei functions}

\item{totals_null}{A boolean. If TRUE, ignore totals_col argument}
}
\description{
Check for missing essential arguments from an ei function
}
\author{
Ari Decter-Frain <agd75@cornell.edu>
}
