% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cor_06}
\alias{cor_06}
\title{Corona 2006 Election Results}
\format{
A data frame with 47 observations on the following 8 variables:
\describe{
 \item{precinct}{Precinct ID number.}
 \item{totvote}{The total vote, per precinct.}
 \item{pct_latino}{Percent of voters identifying as Latino.}
 \item{pct_other}{Percent of voters identifying as non-Latino.}
 \item{pct_breitenbucher}{Percent of vote for Breitenbucher.}
 \item{pct_montanez}{Percent of voters for Montanez.}
 \item{pct_spiegel}{Percent of voters for Spiegel.}
 \item{pct_skipworth}{Percent of voters for Skipworth.}
}
}
\source{
Riverside County, CA Board of Elections.
}
\usage{
data(cor_06)
}
\description{
This dataset contains precinct vote data from a 2006 election in Corona, CA.
}
\keyword{datasets}
