% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corona}
\alias{corona}
\title{Corona 2014 Election Results}
\format{
A data frame with 46 observations on the following 12 variables:
\describe{
 \item{precinct}{Precinct ID number.}
 \item{totvote}{The total vote, per precinct.}
 \item{pct_husted}{Percent of vote for Husted.}
 \item{pct_spiegel}{Percent of vote for Spiegel.}
 \item{pct_ruth}{Percent of vote for Ruth.}
 \item{pct_button}{Percent of vote for Button.}
 \item{pct_montanez}{Percent of vote for Montanez.}
 \item{pct_fox}{Percent of vote for Fox.}
 \item{pct_hisp}{Percent of voters identifying as Hispanic.}
 \item{pct_asian}{Percent of voters identifying as Asian.}
 \item{pct_white}{Percent of voters identifying as white.}
 \item{pct_non_lat}{Percent of voters identifying as non-Latino.}
}
}
\source{
Riverside County, CA Board of Elections
}
\usage{
data(corona)
}
\description{
This dataset contains precinct vote data and racial demographics from a 2014
election in Corona, CA.
}
\keyword{datasets}
