% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_sim.R
\name{.createR}
\alias{.createR}
\title{This function is pulled directly from the ei package to enable a modified
version of ei.sim()}
\usage{
.createR(sub, Rfun, bb, bw, sb, sw, rho, x, numb, numw)
}
\description{
This function is pulled directly from the ei package to enable a modified
version of ei.sim()
}
