% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_sim.R
\name{.samp}
\alias{.samp}
\title{Implement importance sampling}
\usage{
.samp(
  t,
  x,
  n,
  Zb,
  Zw,
  par,
  varcv,
  nsims,
  keep,
  numb,
  covs,
  erho,
  esigma,
  ebeta,
  ealphab,
  ealphaw,
  Rfun
)
}
\description{
This function is pulled directly from the ei package to enable a modified
version of ei.sim()
}
