% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_iter.R
\name{ei_iter}
\alias{ei_iter}
\title{Iterative EI Estimation}
\usage{
ei_iter(
  data,
  cand_cols,
  race_cols,
  totals_col,
  erho = 0.5,
  seed = NULL,
  plots = FALSE,
  betas = FALSE,
  par_compute = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{A data.frame() object containing precinct-level turnout data by
race and candidate}

\item{cand_cols}{A character vector listing the column names for turnout for
each candidate}

\item{race_cols}{A character vector listing the column names for turnout by
race}

\item{totals_col}{The name of the column containing total votes cast in each
precinct}

\item{erho}{A number passed directly to ei::ei(). Defaulted to 0.5}

\item{seed}{A numeric seed value for replicating estimate results across
runs. If NULL, a random seed is chosen. Defaulted to NULL.}

\item{plots}{A boolean indicating whether or not to include density and
tomography plots}

\item{betas}{A boolean to return precinct-level betas for each 2x2 ei}

\item{par_compute}{A boolean to conduct ei using parallel processing}

\item{verbose}{A boolean indicating whether to print out status messages.}

\item{...}{Additional arguments passed directly to ei::ei()}
}
\value{
dataframe of results from iterative ei
}
\description{
This function runs enables running iterative ecoligical inference (EI) to
estimate the proportion of votes by different race/ethnicity groups for
different political candidates.
}
\details{
Iterative EI iterates through all possible race-candidate pairs. For each
pair, votes by other races and for other candidates are binned and 2x2
ecological inference is run.
}
\references{
eiPack. Gary King (1997). A Solution to the Ecological Inference
Problem. Princeton: Princeton University Press.
}
\author{
Loren Collingwood <loren.collingwood@ucr.edu>

Ari Decter-Frain <agd75@cornell.edu>
}
