\name{eiCompare-package}
\alias{eiCompare-package}
\alias{eiCompare}
\docType{package}
\title{
Compares EI, Goodman, RxC Estimates
}
\description{
Compares estimates from three ecological inferences routines, based on King et. al.'s approach.
}
\details{
See demo(demo, "eiCompare") for examples on how to use code
}
\author{
Loren Collingwood

Maintainer: Loren Collingwood <loren.collingwood@ucr.edu>
}
\references{
Gary King (1997). A Solution to the Ecological Inference Problem. Princeton: Princeton University Press.
Lau, Olivia, Ryan Moore, and Michael Kellerman. eiPack: Ecological Inference and Higher-Dimension Data Management
}

\keyword{ package }
