% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ersd_maps}
\alias{ersd_maps}
\title{East Ramapo School District Proposed Maps}
\format{
A data frame with 8 observations on the following 8 variables:
\describe{
 \item{WARD}{The ward ID number.}
 \item{TOT_CVAP}{The total vote, according to CVAP, per precinct.}
 \item{WHI_CVAP}{The number of white CVAP voters, per precinct.}
 \item{BLA_CVAP}{The number of Black CVAP voters, per precinct.}
 \item{HIS_CVAP}{The number of Hispanic CVAP voters, per precinct.}
 \item{ASI_CVAP}{The number of Asian CVAP voters, per precinct.}
 \item{MIN_AGG_FRAC}{The number of Black/Hispanic CVAP voters, per precinct.}
 \item{geometry}{The geomtry for each ward.}
}
}
\source{
East Ramapo School District
}
\usage{
data(ersd_maps)
}
\description{
This dataset contains proposed maps and Citizen Voting Age Population (CVAP)
totals for East Ramapo School District.
}
\keyword{datasets}
