% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwinnett_ei}
\alias{gwinnett_ei}
\title{Stylized dataset of election results and turnout by race in Gwinnett county,
2018 Georgia gubernatorial election.}
\format{
A data frame with 157 rows and 7 columns
\describe{
\item{kemp}{Proportion of votes cast for candidate Brian Kemp}
\item{abrams}{Proportion of votes cast for candidate Stacey Abrams}
\item{metz}{Proportion of votes cast for candidate Jim Metz}
\item{white}{Proportion of voters self-reporting as white}
\item{black}{Proportion of voters self-reporting as black}
\item{other}{Proportion of other voters}
\item{turnout}{Count of voter turnout}
}
}
\usage{
data(gwinnett_ei)
}
\description{
Stylized dataset of election results and turnout by race in Gwinnett county,
2018 Georgia gubernatorial election.
}
\keyword{datasets}
