% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lac_10}
\alias{lac_10}
\title{Los Angeles County 2010 Election Results}
\format{
A data frame with 4980 observations on the following 10 variables:
\describe{
 \item{precinct}{Precinct ID number.}
 \item{tot_reg}{The total number of registered voters.}
 \item{i_jones}{Number of votes for Jones.}
 \item{i_delatore}{Number of votes for Delatorre.}
 \item{votescast}{The total number of votes cast.}
 \item{lat_voters}{Number of Latino voters.}
 \item{pct_latino}{Percent of voters identifying as Latino.}
 \item{pct_delatorre}{Percent of vote for Delatorre.}
 \item{pct_jones}{Percent of vote for Jones.}
 \item{pct_other}{Percent of vote for other candidates.}
}
}
\source{
Los Angeles County
}
\usage{
data(lac_10)
}
\description{
This dataset contains precinct vote data and racial demographics from a 2010
election in Los Angeles County.
}
\keyword{datasets}
