% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong2fips.R
\name{latlong2fips}
\alias{latlong2fips}
\title{Latitude-Longitude Coordinates to FIPS Geocode}
\usage{
latlong2fips(latitude, longitude, number)
}
\arguments{
\item{latitude}{Numeric. Latitude coordinate.}

\item{longitude}{Numeric. Longitude coordinate.}

\item{number}{Numeric. Usually part of a loop index counter}
}
\value{
Character string 15-digit FIPS code corresponding to Lat/Long entry
}
\description{
Converts latitude/longitude coordinates to 15-digit FIPS code. Communicates
with FCC API.
}
\references{
https://geo.fcc.gov/api/census/block/
}
\author{
Loren Collingwood <loren.collingwood@ucr.edu>
}
