% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_sim.R
\name{like}
\alias{like}
\title{Likelihood function for ecological inference.}
\usage{
like(param, y, x, n, Zb, Zw, numb, erho, esigma, ebeta, ealphab, ealphaw, Rfun)
}
\arguments{
\item{param}{passed by .samp}

\item{y}{passed by .samp}

\item{x}{passed by .samp}

\item{n}{passed by .samp}

\item{Zb}{passed by .samp}

\item{Zw}{passed by .samp}

\item{numb}{passed by .samp}

\item{erho}{passed by .samp}

\item{esigma}{passed by .samp}

\item{ebeta}{passed by .samp}

\item{ealphab}{passed by .samp}

\item{ealphaw}{passed by .samp}

\item{Rfun}{passed by .samp#'}
}
\description{
This function is pulled directly from the ei package to enable a modified
version of ei.sim()
}
