% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_ci.R
\name{mean_and_ci}
\alias{mean_and_ci}
\title{mean_and_ci}
\usage{
mean_and_ci(cbind_dat, ci = c(0.025, 0.975))
}
\arguments{
\item{cbind_dat}{cbind object}

\item{ci}{Credible intervals. Default: c(.025, .975)}
}
\value{
Mean and credible interval
}
\description{
Internal
}
\examples{

  
  # EXAMPLE: NOT RUN #
  

}
\author{
Loren Collingwood <loren.collingwood@ucr.edu>
}
