% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiCompare-package.R
\docType{data}
\name{ny_fips}
\alias{ny_fips}
\title{New York state FIPS codes}
\format{
A data frame with 500 observations on the following 2 variables.
\describe{ \item{list("row_id")}{unique id} \item{list("FIP")}{15 digit FIPS
code including state, county, tract, block} }
}
\source{
FCC
}
\description{
New York state FIPS codes example of 500 voters
}
\examples{

data(ny_fips)
str(ny_fips)

}
\keyword{datasets}
