% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ny_fips}
\alias{ny_fips}
\title{New York State FIPS codes}
\format{
A data frame with 500 observations on the following 2 variables:
\describe{
 \item{row_id}{Unique identifier.}
 \item{FIP}{The 15-digit FIPS code.}
}
}
\usage{
data(ny_fips)
}
\description{
New York State FIPS codes for 500 voters.
}
\keyword{datasets}
