% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiCompare-package.R
\docType{data}
\name{ny_voter}
\alias{ny_voter}
\title{New York voter file sample}
\format{
A data frame with 500 observations on the following 10 variables.
\describe{ \item{list("Voter.ID")}{Unique voter file id number, jumbled}
\item{list("SD..Poll")}{Precinct id} \item{list("fips")}{15-digit fips code}
\item{list("st")}{state fips code} \item{list("county")}{county fips code}
\item{list("tract")}{tract fips code} \item{list("block")}{block fips code}
\item{list("st_cty")}{state county fips code}
\item{list("st_cty_tract")}{state county tract fips code}
\item{list("Last.Name")}{Voters' surname} }
}
\source{
ERCSD board of election
}
\description{
New York voter file sample; example of 500 voters
}
\examples{

data(ny_voter)
str(ny_voter)
}
\keyword{datasets}
