% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_utils.R
\name{remove_nas}
\alias{remove_nas}
\title{Remove missing values from dataset and return warning if any removed}
\usage{
remove_nas(data)
}
\arguments{
\item{data}{A dataframe on which ei is to be performed.}
}
\description{
Remove missing values from dataset and return warning if any removed
}
\author{
Ari Decter-Frain <agd75@cornell.edu>
}
