% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_preprocessing.R
\name{standardize_votes}
\alias{standardize_votes}
\title{Computes proportions for returning by clean_race and clean_cand}
\usage{
standardize_votes(votes, new_names = FALSE)
}
\arguments{
\item{votes}{A dataframe of raw votes by race or candidates}

\item{new_names}{A boolean indicating whether to return cand and race columns
with the same names. If FALSE, names returned with "_prop" added on.}
}
\value{
A dataframe of standardized proportions whose columns sum rowwise to
1
}
\description{
Computes proportions for returning by clean_race and clean_cand
}
\author{
Ari Decter-Frain
}
\keyword{internal}
